Implemented rendering order.
Now the elements to render are added to a list where are ordered using its parallax index. A lower index indicates a that must be renderer first, and a higher index to render later, so objects who are nearer to the camera will be renderer later and will be over those with lower parallax index.