Changed rendering order from parallax index to z-index.
Now the rendering order is set by the z-index variable and the parallax index only controls how much the camera movement affects the position of an object.

This abstractions let us to make an static image (parallax index = 0) which can be renderer over all the other objects (high z-index, ie: 99). This can be used for UI elements, for example. It was imposible to archive this with the previous implementation.